remove dependencies on router
