+++ api-lib/frames/MintSuccessFrame.tsx
@@ -1,12 +1,12 @@
 import React from 'react';
 
-import { Frame } from '../../_api/frames/router.tsx';
 import { OGAvatar } from '../../_api/og/OGAvatar.tsx';
 import { insertInteractionEvents } from '../gql/mutations.ts';
 
 import { FramePostInfo } from './_getFramePostInfo.tsx';
 import { getViewerFromParams } from './_getViewerFromParams.ts';
 import { staticResourceIdentifier } from './_staticResourceIdentifier.ts';
+import { Frame } from './frames.ts';
 import { getLevelForViewer } from './give/getLevelForViewer.tsx';
 import {
   FrameBgImage,
