+++ api-lib/frames/_staticResourceIdentifier.ts
@@ -1,4 +1,4 @@
-import { ResourceIdentifier } from '../../_api/frames/router';
+import { ResourceIdentifier } from './frames.ts';
 
 export const staticResourceIdentifier: ResourceIdentifier = {
   resourcePathExpression: '',
