+++ api-lib/frames/giveparty/GivePartyMintWaitingFrame.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 
-import { Frame } from '../../../_api/frames/router.tsx';
 import { OGAvatar } from '../../../_api/og/OGAvatar.tsx';
 import { FramePostInfo } from '../_getFramePostInfo.tsx';
 import { getViewerFromParams } from '../_getViewerFromParams.ts';
+import { Frame } from '../frames.ts';
 import {
   FrameBgImage,
   IMAGE_URL_BASE,
