+++ api-lib/frames/giveparty/JoinedPartyFrame.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 
-import { Frame } from '../../../_api/frames/router.tsx';
 import { fetchPoints } from '../../../_api/hasura/actions/_handlers/createCoLinksGive.ts';
 import { getViewerFromParams } from '../_getViewerFromParams.ts';
+import { Frame } from '../frames.ts';
 import { fetchProfileInfo } from '../give/fetchProfileInfo.tsx';
 import { getLevelForViewer } from '../give/getLevelForViewer.tsx';
 import { FrameBodyGradient } from '../layoutFragments/FrameBodyGradient.tsx';
