+++ api-lib/frames/layoutFragments/FrameBgImage.tsx
@@ -1,7 +1,9 @@
+import { readFileSync } from 'node:fs';
+import { join } from 'node:path';
 import React from 'react';
 
-import { createImage } from '../../../_api/frames/router.tsx';
 import { webAppURL } from '../../../src/config/webAppURL';
+import { IS_LOCAL_ENV } from '../../config.ts';
 
 export const IMAGE_URL_BASE = `${webAppURL('colinks')}/imgs/frames/`;
 
@@ -25,3 +27,17 @@ export const FrameBgImage = ({ src }: { src: string }) => {
     />
   );
 };
+
+export const createImage = (fileNameWithExt: string) => {
+  let imageData: ArrayBuffer;
+  const file = fileNameWithExt.replace('.jpg', '');
+  if (IS_LOCAL_ENV) {
+    imageData = readFileSync(getImagePath(file));
+  } else {
+    imageData = readFileSync(join(__dirname, `./${file}.jpg`));
+  }
+  return imageData;
+};
+
+export const getImagePath = (name: string) =>
+  join(process.cwd(), 'public', 'imgs', 'frames', `${name}.jpg`);
